restructured file for readability and screen format incompatibilities
