--- src/Form.css
@@ -1,65 +1,0 @@
-/*PAGE*/
-.container {
-  height: 100vh;
-  width: 100vw;
-  background-color: #EBF8FA;
-  display: flex;
-  position: relative;
-}
-
-
-/*FORM BACKGROUND*/
-.form-background {
-  height: 90vh;
-  min-width: 600px;
-  background-color: #EA2143;
-  flex-basis: 40%;
-  position: sticky;
-  left: 10vw;
-  top: 5vh;
-  position: relative;
-}
-
-.formText {
-  color: white;
-  font-family: 'Roboto', sans-serif;
-}
-
-.title {
-  font-size: 40px;
-  font-weight: bold;
-  position: absolute;
-  left: 75px;
-  top: 50px;
-}
-
-/*.next {
-  font-size: 18px;
-  color: white;
-  position: absolute;
-  right: 30px;
-  bottom: 20px;
-}
-
-.previous {
-  font-size: 18px;
-  color: white;
-  position: absolute;
-  right: 90px;
-  bottom: 20px;
-}*/
-
-/*IMAGE*/
-.image-container {
-  flex-grow: 1;
-  position: right;
-  height: 100vh;
-  background-color: #EBF8FA;
-  display: flex;
-}
-
-.background-image {
-  height: 100vh;
-  position: sticky;
-  left: 60vw;
-}
