+++ src/components/Stepper.js
@@ -1,9 +1,6 @@
 import React from "react";
 
 const Stepper = ({ currentStep, numberOfStep }) => {
-  console.log('Stepper.currentStep', currentStep)
-  console.log('Stepper.numberOfStep', numberOfStep)
-
   let lines = []
   for(let i = 0 ; i < numberOfStep ; i++) {
     lines.push(
